WIP: Refactor serve() and middleware into new modules
