+++ trustchain-http/src/lib.rs
@@ -6,6 +6,7 @@ pub mod issuer;
 pub mod middleware;
 pub mod qrcode;
 pub mod resolver;
+pub mod server;
 pub mod state;
 pub mod vc;
 pub mod verifier;
